Switch authentication and permission management
Previously, authentication worked by transmitting the password to the
legacy OpenJUB API. Furthermore, user data was also received from this
API and evaluated using a custom JavaScript bridge.

This caused problem, as Jay was able potentially able to intercept user
passwords and also depended on two legacy systems, the OpenJUB API and
the memory intensive JavaScript evaluation of filters

This commit updates the authentication to use the new dreamjub api via
OAuth. Furthermore, it updates the permission system to move away from
UserProfiles and SuperAdmin models. Instead, this commit makes use of
internal Django mechanisms and stores user data directly with the model.

Furthermore, we also migrate the evaluation of filters to use a
python-based evaluation, which no longer needs a JavaScript runtime on
the server side.

See also issue #26.