Do not use JavaScript when evaluating filters
Previously, filters were evaluated using JavaScript. This caused a
performance hit, as it (1) required NodeJS to be installed and (2)
required translation of objects between the languages.

This commit fixes the issue in two steps:
It first removes the simplification of filters, as we can assume this is done by
users manually.
Secondly, it ensures that all evaluation of filters is done using Python.